+++ software/src/soc/STREAM/xrot.h
@@ -20,8 +20,8 @@ namespace coder
   {
     namespace blas
     {
-      void xrot(double x[4], int iy0, double c, double s);
-      void xrot(double x[8], int ix0, int iy0, double c, double s);
+      void xrot(float x[4], int iy0, float c, float s);
+      void xrot(float x[8], int ix0, int iy0, float c, float s);
     }
   }
 }
